Final check - OK

+++ app/Http/Controllers/SiteController.php
@@ -7,6 +7,7 @@ use App\Models\Site;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Requests\SiteRequest;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
 
 class SiteController extends Controller
@@ -19,6 +20,9 @@ class SiteController extends Controller
 
     public function create()
     {
+        $site = new Site();
+        Gate::authorize('crudSite', $site);
+
         return view('site.form',[
             'site' => new Site(),
             'page_meta' => collect([
@@ -31,6 +35,9 @@ class SiteController extends Controller
 
     public function store(SiteRequest $request)
     {
+        $site = new Site();
+        Gate::authorize('crudSite', $site);
+
         $user = User::query()->where('email',$request->email)->first();
         $site = Site::create([
             'name' => $request->validated('name'),
@@ -42,18 +49,23 @@ class SiteController extends Controller
             'description' => $request->validated('description'),
             'user_id' => $user->id??NULL,
         ]);
+
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'created site_id - '.$site->id]);
         return to_route('sites.index');
     }
 
     public function show(Site $site)
     {
+        Gate::authorize('listSite', $site);
+
         $site = Site::query()->where('id',$site->id)->first();
         return view('site.show',['site' => $site]);
     }
 
     public function edit(Site $site)
     {
+        Gate::authorize('crudSite', $site);
+
         return view('site.form',[
             'site' => $site,
             'page_meta' => collect([
@@ -66,6 +78,8 @@ class SiteController extends Controller
 
     public function update(SiteRequest $request, Site $site)
     {
+        Gate::authorize('crudSite', $site);
+
         $user = User::query()->where('email',$request->email)->first();
         $site->update([
             'name' => $request->validated('name'),
@@ -77,14 +91,18 @@ class SiteController extends Controller
             'description' => $request->validated('description'),
             'user_id' => $user->id??NULL,
         ]);
+
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'updated site_id - '.$site->id]);
         return to_route('sites.show',['site' => $site]);
     }
 
     public function destroy(Site $site)
     {
+        Gate::authorize('crudSite', $site);
+
         $site->update(['code' => 'DEL-'.$site->code,'user_id' => NULL]);
         Site::query()->where('id', $site->id)->delete();
+        
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'deleted site_id - '.$site->id]);
         return to_route('sites.index');
     }
